refactor(SiteHeader)
